requirements.txt
# ============================================
# SOCCER ANALYTICS STREAMLIT - TFM
# Multi-Team Analysis Platform
# Compatible with Python 3.10+
# ============================================

# === STREAMLIT CORE ===
streamlit==1.31.0
streamlit-authenticator==0.3.1

# === WEB SCRAPING ===
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# === DATA PROCESSING ===
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5

# === VISUALIZATION ===
plotly==5.24.1

# === IA GENERATIVA ===
# IMPORTANTE: Usar versión 0.28.1 para compatibilidad con la API antigua
openai==0.28.1
python-dotenv==1.0.1

# === PDF GENERATION ===
reportlab==4.2.5
Pillow==11.0.0

# === UTILITIES ===
python-dateutil==2.9.0

# ============================================
# INSTALACIÓN:
# py -m pip install -r requirements.txt
# ============================================
